requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
Django==2.2.2
django-celery-beat==1.5.0
django-celery-results==1.1.0
django-cors-headers==2.4.0
django-timezone-field==3.0
djangorestframework==3.8.2
ecdsa==0.13.2
future==0.17.1
kombu==4.6.3
psycopg2-binary==2.7.5
pyasn1==0.4.5
python-crontab==2.3.6
python-dateutil==2.8.0
python-jose==3.0.1
pytz==2019.1
rsa==4.0
six==1.12.0
sqlparse==0.3.0
Unipath==1.1
vine==1.3.0
